Create reqwest_client module
Split off reqwest implementation for HydraClient into separate module
