+++ src/hydra/mod.rs
@@ -1,2 +1,3 @@
 pub mod client;
+mod reqwest_client;
 mod types;
